EtatDemo

+++ src/main/java/fr/morgan/initiativeasso/service/LikeServiceImpl.java
@@ -2,6 +2,7 @@ package fr.morgan.initiativeasso.service;
 
 import fr.morgan.initiativeasso.model.Like;
 import fr.morgan.initiativeasso.model.User;
+import fr.morgan.initiativeasso.model.exception.ExistingLikeException;
 import fr.morgan.initiativeasso.model.exception.LikeNotFoundException;
 import fr.morgan.initiativeasso.model.exception.UserNotFoundException;
 import fr.morgan.initiativeasso.repository.LikeRepository;
@@ -9,6 +10,7 @@ import fr.morgan.initiativeasso.repository.UserRepository;
 import fr.morgan.initiativeasso.service.interfaces.LikeService;
 
 import java.time.LocalDateTime;
+import java.util.List;
 
 import org.springframework.stereotype.Service;
 
@@ -24,8 +26,10 @@ public class LikeServiceImpl implements LikeService {
     }
 
     @Override
-    public void likeUser(Long userId, Long likedUserId) throws UserNotFoundException {
-
+    public void likeUser(Long userId, Long likedUserId) throws UserNotFoundException, ExistingLikeException {
+        if (likeRepository.existsByUserIdAndLikedUserId(userId, likedUserId))
+            throw ExistingLikeException.builder()
+                    .message("Vous avez deja liké cette personne!").build();
         User user = userRepository.findById(userId).orElseThrow(() -> UserNotFoundException.builder()
                 .message("User not found").build());
 
@@ -42,11 +46,20 @@ public class LikeServiceImpl implements LikeService {
     }
 
     @Override
-    public void unlikeUser(Long userId, Long likedUserId) throws UserNotFoundException, LikeNotFoundException {
-        Like like = likeRepository.findLikeByUserIdAndLikedUserId(userId,likedUserId)
+    public void unlikeUser(Long userId, Long likedUserId) throws UserNotFoundException, LikeNotFoundException, ExistingLikeException {
+        if (!likeRepository.existsByUserIdAndLikedUserId(userId, likedUserId))
+            throw ExistingLikeException.builder()
+                    .message("Vous ne likez pas cette personne!").build();
+
+        Like like = likeRepository.findLikeByUserIdAndLikedUserId(userId, likedUserId)
                 .orElseThrow(() -> LikeNotFoundException.builder().message("Pas de like sur ce user").build());
 
         likeRepository.delete(like);
+    }
+
+    @Override
+    public List<Like> likesListByUserId(Long id) {
+        return likeRepository.findLikesByUserId(id);
     }
 
 }
